Add files via upload
Image steganography is the concealing of information within images in such a way that the viewer believes there is no hidden information.
The idea behind image-based Steganography is very simple. Images are composed of digital data (pixels), which describes what’s inside the picture, usually the colors of all the pixels. Since we know every image is made up of pixels and every pixel contains 3-values (red, green, blue).